third_party/WebKit/LayoutTests/http/tests/devtools/runtime/runtime-callFunctionOn-expected.txt

CONSOLE WARNING: line 3078: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Tests TestRunner.RuntimeAgent.callFunctionOn usages.


Running: testThis
3

Running: testArguments
{"a":1,"b":2}
{"a":1,"b":2}
{"c":1,"d":2}
4
undefined
